myshop/requirements.txt
amqp==5.1.1
asgiref==3.7.2
billiard==4.1.0
Brotli==1.0.9
celery==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.6
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cssselect2==0.7.0
Django==4.2.3
django-debug-toolbar==4.1.0
django-localflavor==4.0
django-parler==2.3
django-rosetta==0.9.9
flower==2.0.0
fonttools==4.41.1
html5lib==1.1
humanize==4.7.0
idna==3.4
kombu==5.3.1
Pillow==10.0.0
polib==1.2.0
prometheus-client==0.17.1
prompt-toolkit==3.0.39
pycparser==2.21
pydyf==0.7.0
pyphen==0.14.0
python-dateutil==2.8.2
python-stdnum==1.18
pytz==2023.3
redis==4.6.0
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
stripe==5.5.0
tinycss2==1.2.1
tornado==6.3.2
tzdata==2023.3
urllib3==2.0.4
vine==5.0.0
wcwidth==0.2.6
weasyprint==59.0
webencodings==0.5.1
zopfli==0.2.2
